Scaffold the articles controller
This commit scaffolds the articles controller using the
`scaffold_controller` rails generator command and creates the articles
model as well. This was done in order to get up and running as quickly
as possible and to get an rspec test suite for articles. Most of these
files will be cleaned up, reorganized, or removed in due time.