Create sessions resource for signing in
This commit creates the sessions resource that includes the controller,
helper, and new view. We include the helper in the application
controller as well. Unnecessary comments in the articles controller were
also cleaned up and the articles index and show views will link to the
creation of a new article if signed in is true.